requirements.txt
Django==1.6.5
South==1.0
argparse==1.2.1
distribute==0.6.24
django-autoslug==1.7.2
django-tastypie==0.11.1
huey==0.4.2
mock==1.0.1
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2014.4
redis==2.10.1
requests==2.3.0
six==1.7.3
wsgiref==0.1.2
